Added REPORT.md & Changes in code

+++ src/Main.java
@@ -2,9 +2,9 @@ import java.util.Arrays;
 
 public class Main {
     public static void main(String[] args) {
-        int current = 2_000_000_000;
-        int replenishment = 500_000_000;
-        int total = current + replenishment;
+        long current = 2_000_000_000;
+        long replenishment = 500_000_000;
+        long total = current + replenishment;
         System.out.println(total);
 
     }
